updated auth service

+++ src/services/auth.service.js
@@ -28,13 +28,18 @@ class AuthService {
     }
 
     logout() {
-        window.localStorage.removeItem("token");
-        window.localStorage.removeItem("user");
-        window.localStorage.removeItem("userId")
-        window.localStorage.removeItem("username");
-        window.localStorage.removeItem("user_roles");
-        window.localStorage.removeItem("user_plain_roles");
-        window.localStorage.clear();
+        return axios.get(API_URL + "/auth/logout").then(res => {
+            if (res.data) {
+                window.localStorage.removeItem("token");
+                window.localStorage.removeItem("user");
+                window.localStorage.removeItem("userId")
+                window.localStorage.removeItem("username");
+                window.localStorage.removeItem("user_roles");
+                window.localStorage.removeItem("user_plain_roles");
+                window.localStorage.clear();
+            }
+            return res.data;
+        })
     }
 
     isAuthenticated() {
